github-notifier: remove "urlencoded" support, remove size restriction
When deploying changes to large static sites, even relatively
small changes where each of the many pages changes a little,
GitHub details the names of every file in the "push" event.

Ref https://github.com/jquery/infrastructure-puppet/issues/10.

These events exceed the size of MAX_BODY_LENGTH. Remove this
restriction but preserve its spirit (not parsing large sequences
of malicious JSON) by validating the HMAC signature *before* parsing
the payload.

For this to be consistent, the support for the "urlencoded" format
is removed because that format required parsing before the "payload"
is known, and thus required parsing before one could validate the
payload. This way, we don't leave the door open to malicious requests
that still invoke `querystring.parse()`.